requirements.txt
google-generativeai
pyTelegramBotAPI
python-dotenv
aiofiles
PyMuPDF
python-docx
python-pptx
pandas
openpyxl
xlrd
duckduckgo-search
ddgs
aiohttp
newspaper3k
beautifulsoup4
Pillow
aiosqlite
lxml